Update content size

+++ mycollab-web/src/main/java/com/esofthead/mycollab/module/ecm/domain/Resource.java
@@ -3,11 +3,11 @@ package com.esofthead.mycollab.module.ecm.domain;
 import java.util.Calendar;
 
 public class Resource {
-	protected String uuid="";
+	protected String uuid = "";
 	protected String createdBy = "";
 	protected Calendar created;
 	protected String path = "";
-	protected long size = 0;
+	protected Double size = 0d;
 	protected String createdUser;
 
 	public String getCreatedBy() {
@@ -34,11 +34,11 @@ public class Resource {
 		return path;
 	}
 
-	public long getSize() {
+	public Double getSize() {
 		return size;
 	}
 
-	public void setSize(long size) {
+	public void setSize(Double size) {
 		this.size = size;
 	}
